# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  filename: string;
  content: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export interface ChatMessage {
  id: number;
  projectId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  files: ProjectFile[];
  messages: ChatMessage[];
}

export interface CreateProjectBody {
  name: string;
  description: string;
}

export interface SendMessageBody {
  content: string;
}

export interface UpdateFileBody {
  content: string;
}

export interface ApiError {
  error: string;
}