# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";
import type { ProjectFile } from "./projectFile";

export interface ProjectDetail {
  id: number;
  name: string;
  description: string;
  createdAt: Date;
  updatedAt: Date;
  files: ProjectFile[];
  messages: ChatMessage[];
}