# lib/api-zod/src/generated/types/projectFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectFile {
  id: number;
  projectId: number;
  filename: string;
  content: string;
  language: string;
  createdAt: Date;
  updatedAt: Date;
}